requirements.txt
absl-py==0.13.0
aiofiles==0.7.0
aioice==0.7.5
aioredis==2.0.0
aiortc==1.2.1
aiosqlite==0.17.0
alembic==1.7.5
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.24.0
atomicwrites==1.4.0
attrs==21.2.0
av==8.0.3
Babel==2.9.1
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
cycler==0.10.0
databases==0.4.3
Deprecated==1.2.12
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.0
fastapi-admin==1.0.3
fastapi-users==6.1.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-crc32c==1.1.2
google-pasta==0.2.0
grpcio==1.39.0
h11==0.12.0
h5py==3.1.0
httptools==0.2.0
idna==3.2
imageai==2.1.5
iniconfig==1.1.1
iso8601==0.1.16
Jinja2==3.0.1
jose==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
makefun==1.11.3
Mako==1.1.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
multidict==5.1.0
netifaces==0.10.9
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
ormar==0.10.16
packaging==21.3
passlib==1.7.4
pendulum==2.1.2
Pillow==8.3.1
pip==21.2.4
pluggy==1.0.0
protobuf==3.17.3
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
pydantic==1.8.2
pyee==8.2.2
PyJWT==2.1.0
pylibsrtp==0.6.8
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
setuptools==58.0.4
six==1.16.0
SQLAlchemy==1.3.23
starlette==0.14.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
toml==0.10.2
tortoise-orm==0.17.8
typing-extensions==3.7.4.3
urllib3==1.26.6
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
Werkzeug==2.0.1
wheel==0.37.0
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
